move fastani module to separate file
